Bumped spec version to 5 with many changes
This addresses a lot of the feedback from the past couple of months.

- Changed `VkPresentTimesInfoEXT` and `VkPresentTimeInfoEXT` to
`VkPresentTimingsInfoEXT` and `VkPresentTimingInfoEXT` because those structs
contain more than just time values now.

- Remove `presentStageQueries` and `presentStageTarget` from
VkSwapchainPresentTimingCreateInfoEXT. These are specified per-present now, to
accomodate for dynamic surface capabilities.

- Remove vkSetSwapchainTimingEXT function and VkSwapchainTimingInfoEXT
structure. The state specified there is now sent per-present.

- Add a synchronous wait behavior to vkGetPastPresentationTimingEXT with a
`timeout` parameter, and the ability to explicitly reference present IDs as an
input.  vkGetPastPresentationTimingEXT also does not require external
synchronization anymore.

- Remove `timingPropertiesChanged` from `VkPastPresentationTimingEXT` struct
and make it an out parameter of vkGetPastPresentationTimingEXT. This means
applications don't need to wait for timing results to be available to know that
timing properties have changed.  This however creates a new problem where if
multiple results are returned, applications cannot distinguish the timing
properties of individual presents.

- Remove `timeDomainChanged` from `VkPastPresentationTimingEXT` struct and
replace it with the actual time domain enum it was changed to. The PE can
operate some fallback and communicate this to the app for each result. This
also creates a new problem if both the old and new time domains are the same
opaque one, e.g. different SWAPCHAIN_LOCAL domains. I have plans to address
this and the timing properties problem, but would like to discuss it in future
meetings first.

- Fix relative present timing language to correctly specify that the duration
is relative to the previous image, and not the next present.

And many other minor updates