Remove timeout from vkGetPastPresentationTiming
This change mostly removes the ability to use vkGetPastPresentationTiming synchronously
with a timeout. Not all implementors can support this, and it wouldn't make sense
to have it as a feature bit.

This also incorporates the latest feedback regarding timing properties changes, which
are now tracked through a counter exposed to the application.

Other minor fixes and cleanup.